Removed public pivot API.
Moved pivot assertions into unit test
